WEB: Hide demos with the same target IDs by default (can be toggled)
Now, only the first demo will be shown from a group of demos with the
same target ID. Icons are shown next to each grouped download, which
allow the users to open/close the groups (forming virtual "folders").
This simplifies the demos page a lot, without any changes to the
existing data. This feature is a progressive enhancement: the demos
page will still look the same when Javascript is disabled