WEB: Update compatibility listings for SCI32 games where appropriate
Games which have been tested and which do not use MIDI or have
pathfinding bugs are upgraded to 'excellent' support level, and
platform names are removed for games with only a single platform
release (i.e. when a game came with DOS & Windows interpreters but
there was no difference between the two so we only detect them as
having a single OS option).

QFG4 is marked as broken instead of untested, because it is
currently pretty broken.