NEWS: Add support for news filenames in either yyyyMMdd or yyyyMMddHHmm format
This also changes the behavior that when a user clicks on a news
article, he will get all news articles from the same day. The reasoning
is a) we barely have more than 1 article a day. b) if a user wanted a
hotlink to a particular news article, he might get two different ones
which is odd.