WEB: Use macOS instead of Mac OS X in supported platform list
The Mac OS name has changed several times. Mac OS X was the original
name before it was changed to OS X with the release of OS X 10.8,
and then to macOS with the release of macOS 10.12. Using the current
name seems better, even though we support older versions of the
system.