WEB: Add a Wikipedia link for games
Add a new field to the game table called `wikipedia_page`, which is the URL page title for the game's corresponding page on English Wikipedia.

The field has been filed out for a few games (such as Martian Memorandum and Manhunter: New York), but more will need to be filled out.