WEB: Removing collapsible demo lists
The current system of collapsible demo lists has a number of problems:

* It is inconvenient to use on mobile
* It hinders searching for text on the page (e.g. Ctrl+F and searching for "French" will not find all the French demos)
* The first demo (e.g. English Amiga) may have nothing to do with the other demos (e.g. French DOS), yet it is the first one visible

Therefore, I am removing the entirety of the collapsible lists to avoid these issues. The main downside is that the demo pages are now about double the length.

Perhaps we could find a new alternative method of organizing that does not suffer from these problems. Until then, this will allow everyone on mobile and desktop to easily find all the demos they might be looking for.