WEB: Removing SCUMM hacking forum link and moving AGI Development Site (#243)
The SCUMM hacking forum is a sub-site of MixNMojo, which is already listed, and it's been archived for six years. I don't think we need a separate link.

Then with the AGI site being the only one in the category, I'm merging it with the larger category of external sites.