IMAGES: Replace Blue Force screenshots with updated ones
These screenshots were taken after the recent bug fixes for Tsage and Blueforce

Screenshots were chosen so that they're as close as possible to what was depicted in the old ones.
The most noticeable difference is in blueforce_dos_en_1_7_full.png where the proper face image is used for Laura (as in the original).

Since the game is 320x200 the screenshots were taken with a scaler Normal x2, pixel perfect scaling for stretch mode, enabled Aspect ratio correction and SDL Surface for graphics mode.
The resulting pngs were run through OptiPNG 7.9.1-win64 (default arguments)